\name{gsub_eval_opt}
\alias{gsub_eval_opt}
\title{
Parameters for the circular layout
}
\description{
Parameters for the circular layout
}
\usage{
gsub_eval_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{

  \item{...}{Arguments for the parameters, see "details" section}
  \item{RESET}{reset to default values}
  \item{READ.ONLY}{please ignore}
  \item{LOCAL}{please ignore}
  \item{ADD}{please ignore}

}
\details{
There are following global options:

\describe{
  \item{\code{engine}}{which function to do variable intepolation. By default it is \code{\link[GetoptLong]{qq}}.}
}
}
\examples{
# There is no example
NULL
}
